Monitor magic tests as mdx
In particular this allows to show that with the changes to make the
library Type-Safe, the definition of such trait is no longer possible
by design.

This is possible to show using mdx files since we can capture type
errors in expect traces, whereas in the test/ we simply have to delete
these tests.